fix(IdentificationNumber): set user IdentificationNumber as long

+++ ErgoProxy.Domain/Document/AuthenticateDocumentDto.cs
@@ -5,7 +5,7 @@ using System.Text.Json.Serialization;
 public class AuthenticateDocumentDto
 {
     [JsonPropertyName("idCitizen")]
-    public int IdCitizen {  get; set; }
+    public long IdCitizen { get; set; }
     [JsonPropertyName("UrlDocument")]
     public string UrlDocument { get; set; }
     [JsonPropertyName("documentTitle")]
